Skip weekend

+++ apocarich/data.py
@@ -1,4 +1,5 @@
 import datetime
+import glob
 import subprocess
 from pathlib import Path
 
@@ -7,6 +8,30 @@ import os
 import seaborn as sns
 
 sns.set(color_codes=True)
+
+DATE_FORMAT = "%Y-%m-%d"
+
+
+def is_weekend(date_string):
+    date = datetime.datetime.strptime(date_string, DATE_FORMAT)
+    return date.weekday() >= 5
+
+
+def is_today(date_string):
+    return datetime.datetime.now().strftime(DATE_FORMAT) == date_string
+
+
+def create_empty_file(path):
+    with open(path, "w"):
+        pass
+
+
+def remove_all_may_be_incomplete_files():
+    for filename in glob.iglob("data/**", recursive=True):
+        if os.path.isfile(filename):  # filter dirs
+            if "MAY_BE_INCOMPLETE" in filename:
+                print(f"Removing {filename}...")
+                Path(filename).unlink()
 
 
 def read_data_directory(
@@ -32,16 +57,22 @@ def read_data_directory(
     )
 
 
-def retrieve_aws_data(trading_platform="xetra", date="2017-06-17", skip_duplicate=True):
+def retrieve_aws_data(
+    date, trading_platform="xetra", skip_duplicate=True, skip_weekend=True
+):
+    if skip_weekend and is_weekend(date):
+        print(f"[SKIP]\tSkipping weekend day {date}...")
+        return
+
     target_path = Path("data", f"deutsche-boerse-{trading_platform}-pds", date)
     files = sorted([f.stem for f in list(target_path.glob("*.csv"))])
 
     if skip_duplicate and len(files) > 0:
         if Path(target_path, "MAY_BE_INCOMPLETE").exists():
-            print("Found possibly incomplete data. Renewing..")
+            print(f"[?]\t{date}\tFound possibly incomplete data. Renewing...")
             Path(target_path, "MAY_BE_INCOMPLETE").unlink()
         else:
-            print(f"Data for {date} already exists. Skipping...")
+            print(f"[SKIP]\tData for {date} already exists. Skipping...")
             return
 
     command = [
@@ -54,32 +85,29 @@ def retrieve_aws_data(trading_platform="xetra", date="2017-06-17", skip_duplicat
     ]
     subprocess.run(command, stdout=subprocess.PIPE, universal_newlines=True)
 
-    date_format = "%Y-%m-%d"
-    if datetime.datetime.now().strftime(date_format) == "2020-03-25":
-        with open(Path(target_path, "MAY_BE_INCOMPLETE"), "w"):
-            pass
+    if is_today(date):
+        create_empty_file(Path(target_path, "MAY_BE_INCOMPLETE"))
+
+    print(f"[OK]\t{date}")
 
 
-def retrieve_all_aws_data(trading_platform="xetra", start_date="2019-12-01"):
+def retrieve_all_aws_data(start_date="2019-12-01", trading_platform="xetra"):
     """
     :param trading_platform: "xetra" or "eurex"
     :param start_date: Earliest possible is "2017-06-17" for Xetra AND "2017-05-27" for Eurex
     :return:
     """
-    date_format = "%Y-%m-%d"
-    # start = datetime.datetime.strptime("2017-06-17", date_format)
-    start = datetime.datetime.strptime(start_date, date_format)
-    # end = datetime.datetime.strptime("2017-06-19", date_format)
+    start = datetime.datetime.strptime(start_date, DATE_FORMAT)
     end = datetime.datetime.now()
 
     generated_dates = [
         start + datetime.timedelta(days=x) for x in range(0, (end - start).days + 1)
     ]
-    dates = [date.strftime(date_format) for date in generated_dates]
+    dates = [date.strftime(DATE_FORMAT) for date in generated_dates]
 
     for date in dates:
         print(f"Retrieving stock data form {trading_platform} for day {date}...")
-        retrieve_aws_data(trading_platform, date)
+        retrieve_aws_data(date, trading_platform)
 
 
 def filter_common_stocks(dataframe):
@@ -89,3 +117,8 @@ def filter_common_stocks(dataframe):
 def do():
     df = read_data_directory("2019-12-02").pipe(filter_common_stocks)
     df["price"] = df.StartPrice + df.EndPrice / 2
+
+
+def main():
+    start_date = "2019-11-01"
+    retrieve_all_aws_data(start_date=start_date)
